requirements.txt
pywebview>=5.0
sounddevice>=0.5
numpy>=1.26
httpx>=0.27
python-dotenv>=1.0
speechbrain>=1.0
torch>=2.0
torchaudio>=2.0
